feat(i2c_slave): Implement i2c_slave driver callback based for esp32c6
The driver is only tested on esp32c6.
It relies on clock stretching and the stretch interrupt.
The interface to an application is via on "on_events" callback
that the application can register to.
The callback will be directly called from the i2c interrupt and must not block!

Tests on esp32c6 show low latency (e.g. typical clock stretch delays around 40us.
It can handle write/read transactions from master with various lengths.